refactor(-): :hammer: move viewsets logs dietas
