+++ sme_sigpae_api/dieta_especial/urls.py
@@ -1,6 +1,10 @@
 from django.urls import include, path, re_path
 from rest_framework import routers
 
+from sme_sigpae_api.dieta_especial.logs_models.api import (
+    viewsets as logs_models_viewsets,
+)
+
 from .api import viewsets
 from .constants import (
     ENDPOINT_ALERGIAS_INTOLERANCIAS,
@@ -54,22 +58,22 @@ router.register(
 )
 router.register(
     "log-quantidade-dietas-autorizadas",
-    viewsets.LogQuantidadeDietasAutorizadasViewSet,
+    logs_models_viewsets.LogQuantidadeDietasAutorizadasViewSet,
     basename="Log quantidade dietas autorizadas",
 )
 router.register(
     "log-quantidade-dietas-autorizadas-cei",
-    viewsets.LogQuantidadeDietasAutorizadasCEIViewSet,
+    logs_models_viewsets.LogQuantidadeDietasAutorizadasCEIViewSet,
     basename="Log quantidade dietas autorizadas CEI",
 )
 router.register(
     "log-quantidade-dietas-autorizadas-recreio-nas-ferias",
-    viewsets.LogQuantidadeDietasAutorizadasRecreioNasFeriasViewSet,
+    logs_models_viewsets.LogQuantidadeDietasAutorizadasRecreioNasFeriasViewSet,
     basename="Log quantidade dietas autorizadas RecreioNasFerias",
 )
 router.register(
     "log-quantidade-dietas-autorizadas-recreio-nas-ferias-cei",
-    viewsets.LogQuantidadeDietasAutorizadasRecreioNasFeriasCEIViewSet,
+    logs_models_viewsets.LogQuantidadeDietasAutorizadasRecreioNasFeriasCEIViewSet,
     basename="Log quantidade dietas autorizadas RecreioNasFerias CEI",
 )
